Changed syntax to specify ranges like field lengths or rules for for Integer fields formats. Use ":" instead of "...".
There are basically two reasons for this change: Firstly, this looks more pythonic and thus more consistent with other parts of the ICD like the "Checks" section which also uses Python syntax in various places. Secondly, this avoids issues with Excel which under certain circumstances changes the 3 characters in "..." to a single character ellipsis. Using ":" still is not without issues though: if you use a spreadsheet application to author ICDs, most of them think of a value like "1:60" (which could for example specify a field length between 1 and 60 characters) to refer to a time of 1 hour and 60 minutes. To avoid any confusion, disable the cell format auto detection of the spreadsheet application by changing all cells to contain "Text".