Fixed checks when validating more than one data file from the command line.
Until now the checks did preserve internal state information needed to perform the check. For instance, !IsUniqueCheck remembered the keys of all rows read so far. So when a data file contained a row with a key that already showed up in an earlier data file, the check failed. To prevent this from happening, validate() now resets all checks. See also: Ticket #9.