#74: Cleanup project structure and build process.
Changed performance test to process only 100 instead of 20000 rows in order to speed up the tests. Before #74, `test_performance` was not part of the standard test suit and run separately by Jenkins CI to collect a performance graph over time. With the move to Travis CI using pyscaffold's default setup, this functionality seems to be gone, so for the time being the performance test is of rather limited use. Nevertheless we should strive to eventually bring it back as a separate build target for Travis CI.