#56: Add support for Python 3.
Changed explicit relative imports back to absolute imports - again, and this time for good. Rationale: absolute imports are the default with Python 3 and explicit relative import have been discouraged by PEP8 for a while. More recent versions of PEP8 describe a few scenarios where explicit relative imports might be useful, none of them apply for cutplace.

The main drawback of explicit relative imports is that they are almost impossible to work if the same source code should support

* python cutplace/xxx.py
* python -m cutplace.xxx
* python setup.py test
* pytest cutplace/xxx.py
* run and debug from an IDE like PyCharm or PyDev

The only drawback seems to be that pylint thinks that `from cutplace import _tools` is bad while `from . import _tools` is fine but I consider this a bug in pylint.