#66: Cleanup PEP8 issues.
Cleaned up indentation of long parameter lists, deliberately ignoring the (weird) PEP8 recommendation to write as many characters as possible and then continue at the same column as the "(" in the line above. As this constantly breaks the indentation after renames, there does not seem to be much point to it.